minorUpdate

+++ process_multiModels_multiTrials.m
@@ -16,26 +16,28 @@ clear;
 import org.opensim.modeling.*;
 
 
-setupIKfile = fullfile(pwd, 'SetupFiles', 'setupIK_CP.xml');
-setupIK_reducedTorso_file = fullfile(pwd, 'SetupFiles', 'setupIK_CP_reducedTorso.xml');
-% setupIKfile = fullfile(pwd, 'SetupFiles', 'setupIK_DEM_TD.xml');
-% setupIK_reducedTorso_file = fullfile(pwd, 'SetupFiles', 'setupIK_DEM_TD_reducedTorso.xml');
-% setupIKfile = 'C:\Users\Willi\ucloud\ProjekteAbteilung\Climbing\OpenSimSetup\ik_setup4.xml';
-% setupIK_reducedTorso_file = 'C:\Users\Willi\ucloud\ProjekteAbteilung\Climbing\OpenSimSetup\ik_setup4.xml';
+% setupIKfile = fullfile(pwd, 'SetupFiles', 'setupIK_CP.xml');
+% setupIK_reducedTorso_file = fullfile(pwd, 'SetupFiles', 'setupIK_CP_reducedTorso.xml');
+setupIKfile = fullfile(pwd, 'SetupFiles', 'Settings_IK_plugInGait.xml');
+setupIK_reducedTorso_file = []; %fullfile(pwd, 'SetupFiles', 'Settings_IK_plugInGait.xml');
 setupIDfile = fullfile(pwd, 'SetupFiles', 'Settings_ID.xml');
 setupSOfile = fullfile(pwd, 'SetupFiles', 'Settings_SO.xml');
-% setupJRLfile = fullfile(pwd, 'SetupFiles', 'Settings_JRL.xml');
-setupJRLfile = fullfile(pwd, 'SetupFiles', 'Settings_JRL_inFemurL_with_MuscleDirection_BodyKinematics.xml');
-setupJRL2file = fullfile(pwd, 'SetupFiles', 'Settings_JRL_inFemurR_with_MuscleDirection_BodyKinematics.xml');
+setupJRLfile = fullfile(pwd, 'SetupFiles', 'Settings_JRL.xml');
+% setupJRLfile = fullfile(pwd, 'SetupFiles', 'Settings_JRL_inFemurL_with_MuscleDirection_BodyKinematics.xml');
+% setupJRL2file = fullfile(pwd, 'SetupFiles', 'Settings_JRL_inFemurR_with_MuscleDirection_BodyKinematics.xml');
 
-% Options for steps which should be run
+%% Options for steps which should be run
 b_runIK = 1;
-b_checkMuscleMomentArms = 0; 
+b_checkMuscleMomentArms = 0; % there is a better script to check for discontinuities in muscle moment arms - https://doi.org/10.1016/j.gaitpost.2025.01.063
+% best is to run only IK in the first step;
+% then, use the other script to check and resolve muscle discontinuities;
+% then, run ID / SO / JRL 
 b_skipTrialIfMomentArmsWrong = 0;
-b_runID = 0;
-b_runSO = 0;
+b_runID = 1;
+b_runSO = 1;
 b_runJRL = 1;
 
+%%
 disp('Select the root output folder');
 rootOutput = uigetdir(pwd, 'Select the root output folder');
 if isequal(rootOutput, 0)
@@ -88,6 +90,7 @@ for trialIndex = 1 : size(trialsFileNames, 2)
         startTime = 0;
         endTime = duration;
     end
+
     
     preframes = startTime * frequency;
 
@@ -126,19 +129,21 @@ for trialIndex = 1 : size(trialsFileNames, 2)
                 % first run with reduced markers for torso to get
                 % interesting marker errors for lower legs
 
-                disp('without torso ... ');
-                resultsAreValid = runIK(setupIK_reducedTorso_file, modelFile, trcFile, startTime, endTime, fullfile(outputPath, 'IK_reducedTorsoMarkers'), 1, frequency);
-                if(~resultsAreValid)
-                    fprintf(2,'Maker Error too big for inverse Kinematics!\n'); 
+                if ~isempty(setupIK_reducedTorso_file)
+                    disp('without torso ... ');
+                    resultsAreValid = runIK(setupIK_reducedTorso_file, modelFile, trcFile, startTime, endTime, fullfile(outputPath, 'IK_reducedTorsoMarkers'), 1, frequency);
+                    if(~resultsAreValid)
+                        fprintf(2,'Maker Error too big for inverse Kinematics!\n');
 
-                    fileID = fopen(fullfile(rootOutput, modelFileNameNoExt, trialName, 'IK_fehler.txt'),'w');
-                    fprintf(fileID,'Marker errors higher than recommended! Check again!\n');
-                    fclose(fileID);
+                        fileID = fopen(fullfile(rootOutput, modelFileNameNoExt, trialName, 'IK_fehler.txt'),'w');
+                        fprintf(fileID,'Marker errors higher than recommended! Check again!\n');
+                        fclose(fileID);
+                    end
+
+                    % now run the same with more torso markers
+                    % This overwrites the output from the previous run
+                    disp('and now with torso ... ');
                 end
-
-                % now run the same with more torso markers
-                % This overwrites the output from the previous run
-                disp('and now with torso ... ');
                 resultsAreValid = runIK(setupIKfile, modelFile, trcFile, startTime, endTime, fullfile(outputPath, 'IK'), 0, frequency);
                 if(~resultsAreValid)
                     fileID = fopen(fullfile(rootOutput, modelFileNameNoExt, trialName, 'IK_withTorsoOutOfRecommendation.txt'),'w');
@@ -175,7 +180,7 @@ for trialIndex = 1 : size(trialsFileNames, 2)
                 disp('Estimating Joint Contact Forces ... ');
                 so_forcesFile = fullfile(outputPath, 'SO', '_StaticOptimization_force.sto');
                 runJRL(setupJRLfile, actuatorfile, modelFile, motionFile, grfSetupFile, so_forcesFile, startTime, endTime, fullfile(outputPath, 'JRL'));
-                if isfile(setupJRL2file)
+                if exist('setupJRL2file', 'var') && isfile(setupJRL2file)
                     copyfile(fullfile(outputPath, 'JRL', '_JointReaction_ReactionLoads.sto'), fullfile(outputPath, 'JRL', 'inFemurL_JointReaction_ReactionLoads.sto'));
                     runJRL(setupJRL2file, actuatorfile, modelFile, motionFile, grfSetupFile, so_forcesFile, startTime, endTime, fullfile(outputPath, 'JRL'));
                     copyfile(fullfile(outputPath, 'JRL', '_JointReaction_ReactionLoads.sto'), fullfile(outputPath, 'JRL', 'inFemurR_JointReaction_ReactionLoads.sto'));
